Sensors completed

+++ Implementation/Utility/RandomNumberGenerator/random-number-generator.c
@@ -4,7 +4,7 @@
 double init_random_number(double min, double max){
     
     // Generating a random number between 0 and 1
-    double random_value = random_rand() / RANDOM_RAND_MAX;
+    double random_value =  (double) random_rand() / RANDOM_RAND_MAX;
 
     // Scale and shift to the desired range [min_value, max_value]
     return min + (max - min) * random_value;
@@ -15,7 +15,7 @@ double generate_random_number(double min, double max, double old_value, double m
     double mean = (min + max) / 2;
     double max_variation = max_delta * mean;      
 
-    double variation = (random_rand() / RANDOM_RAND_MAX) * max_variation;   // value between 0 and max_variation
+    double variation = ((double) random_rand() / RANDOM_RAND_MAX) * max_variation;   // value between 0 and max_variation
     
 
     if(hvac_status){
